backend/requirements.txt


Django~=5.0.3
djangorestframework~=3.16.1

psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.28
openai>=0.28.0
django-cors-headers>=3.13
